Consolidated header files.
Saw another code base where several C files of similar purpose shared headers,
I really liked it and thought it would be nice to get rid of some files and
reduced the amount of #include's I had in each C file.

Also, moved the percentage label into the top right of the CPU and Memory windows.
