+++ src/include/startup.h
@@ -3,7 +3,7 @@
 
 #include <pthread.h>
 
-#include "../monitor/proc_monitor.h"
+#include "monitor.h"
 
 extern volatile PROC_STATS **procStats;
