+++ src/include/thread_safe_queue.h
@@ -1,5 +1,5 @@
-#ifndef SHARED_QUEUE_H
-#define SHARED_QUEUE_H
+#ifndef THREAD_SAFE_QUEUE_H
+#define THREAD_SAFE_QUEUE_H
 
 #include <pthread.h>
 
@@ -13,32 +13,32 @@ typedef struct _shared_queue
 {
 	QUEUE_NODE *head, *tail;
 	int size;
-} SHARED_QUEUE;
+} ThreadSafeQueue;
 
 void enqueue(
-	SHARED_QUEUE *q,
+	ThreadSafeQueue *q,
 	void *stats,
 	pthread_mutex_t *queueLock,
 	pthread_cond_t *condition
 );
 int dequeue(
-	SHARED_QUEUE *q,
+	ThreadSafeQueue *q,
 	pthread_mutex_t *queueLock,
 	pthread_cond_t *condition
 );
 void * peek(
-	SHARED_QUEUE *q,
+	ThreadSafeQueue *q,
 	pthread_mutex_t *queueLock,
 	pthread_cond_t *condition
 );
 int timedDequeue(
-	SHARED_QUEUE *q,
+	ThreadSafeQueue *q,
 	pthread_mutex_t *queueLock,
 	pthread_cond_t *condition,
 	unsigned short timeoutSec
 );
 void * timedPeek(
-	SHARED_QUEUE *q,
+	ThreadSafeQueue *q,
 	pthread_mutex_t *queueLock,
 	pthread_cond_t *condition,
 	unsigned short timeoutSec
