+++ src/thread/thread.c
@@ -1,4 +1,4 @@
-#include "../include/thread/thread.h"
+#include "../include/thread.h"
 
 pthread_mutex_t cpuQueueLock;
 pthread_mutex_t memQueueLock;
